Update: split html and install html loader
add separate html for task and project dialog
