+++ 11_To_Do/src/components/html/project-modal.html
@@ -1,0 +1,24 @@
+<dialog class="project-dialog">
+  <form method="dialog" id="task-form" action="">
+    <div class="task-title">
+      <label for="task-title">Enter task title: </label>
+      <input
+        type="text"
+        name="task-title"
+        id="task-title"
+        required
+        value="Task 1"
+      />
+    </div>
+    <div class="task-description">
+      <label for="task-description">Enter task description: </label>
+      <input
+        type="text"
+        name="task-description"
+        id="task-description"
+        required
+        value="Description 1"
+      />
+    </div>
+  </form>
+</dialog>
