+++ 11_To_Do/src/index.html
@@ -7,80 +7,11 @@
   </head>
   <body>
     <h1>To Do App</h1>
-    <div class="project" id="project1">
+    <div class="project">
       <p>Project 1</p>
-      <dialog class="task-dialog" data-project="Project 1">
-        <form method="dialog" id="task-form" action="">
-          <div class="task-title">
-            <label for="task-title">Enter task title: </label>
-            <input
-              type="text"
-              name="task-title"
-              id="task-title"
-              required
-              value="Task 1"
-            />
-          </div>
-          <div class="task-description">
-            <label for="task-description">Enter task description: </label>
-            <input
-              type="text"
-              name="task-description"
-              id="task-description"
-              required
-              value="Description 1"
-            />
-          </div>
-          <div class="task-due">
-            <label for="task-due">Select task due date: </label>
-            <input
-              type="date"
-              name="task-due"
-              id="task-due"
-              required
-              value="2024-05-05"
-            />
-          </div>
-          <div class="task-priority">
-            <label for="task-priority">Select task importance: </label>
-            <div>
-              <input
-                type="radio"
-                id="low"
-                name="task-priority"
-                value="low"
-                checked
-              />
-              <label for="low">LOW</label>
-            </div>
-            <div>
-              <input
-                type="radio"
-                id="medium"
-                name="task-priority"
-                value="medium"
-              />
-              <label for="medium">MEDIUM</label>
-            </div>
-            <div>
-              <input type="radio" id="high" name="task-priority" value="high" />
-              <label for="high">HIGH</label>
-            </div>
-          </div>
-          <div class="task-project">
-            <label for="task-project">For project: </label>
-            <input
-              type="text"
-              name="task-project"
-              id="task-project"
-              required
-              value="Project 1"
-            />
-          </div>
-          <button type="submit">Add Task</button>
-        </form>
-      </dialog>
-      <button class="create-task">Add New Task</button>
+      <button class="create-task" id="project1">Add New Task</button>
     </div>
+    <div id="taskDialogHTML"></div>
+    <div id="projectDialogHTML"></div>
   </body>
 </html>
