+++ 11_To_Do/webpack.config.js
@@ -10,12 +10,15 @@ module.exports = {
   },
   plugins: [
     new HtmlWebpackPlugin({
-      title: "Output Management",
       template: "./src/index.html",
     }),
   ],
   module: {
     rules: [
+      {
+        test: /\.html$/,
+        use: "html-loader",
+      },
       {
         test: /\.css$/i,
         use: ["style-loader", "css-loader"],
